// web/graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Node = {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Internal = {
  content?: Maybe<Scalars['String']>;
  contentDigest: Scalars['String'];
  description?: Maybe<Scalars['String']>;
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
  ignoreType?: Maybe<Scalars['Boolean']>;
  mediaType?: Maybe<Scalars['String']>;
  owner: Scalars['String'];
  type: Scalars['String'];
};
// ...
export type StrapiProduct = Node & {
  interval?: Maybe<Scalars['String']>;
  bonus_tickets?: Maybe<Scalars['Boolean']>;
  display_price?: Maybe<Scalars['String']>;
  credits_awarded?: Maybe<Scalars['Int']>;
  credit_cost?: Maybe<Scalars['Int']>;
  referral_multiplier?: Maybe<Scalars['Float']>;
  stripe_plan_id?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  amount?: Maybe<Scalars['Int']>;
  type?: Maybe<Scalars['String']>;
  display_name?: Maybe<Scalars['String']>;
  isActive?: Maybe<Scalars['Boolean']>;
  order?: Maybe<Scalars['Int']>;
  isPrepaidAnnualPlan?: Maybe<Scalars['Boolean']>;
  inboundcampaign?: Maybe<StrapiProductInboundcampaign>;
  outboundcampaigns?: Maybe<StrapiProductOutboundcampaigns>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  badge?: Maybe<StrapiProductBadge>;
  dynamic_pricing?: Maybe<StrapiProductDynamic_Pricing>;
  strapiId?: Maybe<Scalars['Int']>;
  id: Scalars['ID'];
  giftcode?: Maybe<Scalars['Int']>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type StrapiProductInboundcampaign = {
  id?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  campaign_id?: Maybe<Scalars['String']>;
  super_hero_content?: Maybe<Scalars['String']>;
  isActive?: Maybe<Scalars['Boolean']>;
  visitors?: Maybe<Scalars['Int']>;
  confirmed_signups?: Maybe<Scalars['Int']>;
  cta_button_label?: Maybe<Scalars['String']>;
  public_url?: Maybe<Scalars['String']>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
};
// ...
export type StrapiProductOutboundcampaigns = {
  id?: Maybe<Scalars['Int']>;
  headline?: Maybe<Scalars['String']>;
  headline_2?: Maybe<Scalars['String']>;
  subtitle?: Maybe<Scalars['String']>;
  body_content?: Maybe<Scalars['String']>;
  cta_button_label?: Maybe<Scalars['String']>;
  campaign_id?: Maybe<Scalars['String']>;
  public_url?: Maybe<Scalars['String']>;
  isActive?: Maybe<Scalars['Boolean']>;
  style?: Maybe<Scalars['String']>;
  product?: Maybe<Scalars['Int']>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
};
// ...
export type StrapiProductBadge = {
  id?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  is_public?: Maybe<Scalars['Boolean']>;
  icon?: Maybe<StrapiProductBadgeIcon>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
  product?: Maybe<Scalars['Int']>;
};
// ...
export type StrapiProductBadgeIcon = {
  id?: Maybe<Scalars['Int']>;
  name?: Maybe<Scalars['String']>;
  alternativeText?: Maybe<Scalars['String']>;
  caption?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
  formats?: Maybe<StrapiProductBadgeIconFormats>;
  hash?: Maybe<Scalars['String']>;
  ext?: Maybe<Scalars['String']>;
  mime?: Maybe<Scalars['String']>;
  size?: Maybe<Scalars['Float']>;
  url?: Maybe<Scalars['String']>;
  provider?: Maybe<Scalars['String']>;
  provider_metadata?: Maybe<StrapiProductBadgeIconProvider_Metadata>;
  created_at?: Maybe<Scalars['Date']>;
  updated_at?: Maybe<Scalars['Date']>;
};
// ...
export type StrapiProductDynamic_Pricing = {
  id?: Maybe<Scalars['Int']>;
  gold_price?: Maybe<Scalars['Int']>;
  gold_per_credit_rate?: Maybe<Scalars['Float']>;
  platinum_price?: Maybe<Scalars['Int']>;
  platinum_per_credit_rate?: Maybe<Scalars['Float']>;
  founders_price?: Maybe<Scalars['Int']>;
  founders_per_credit_rate?: Maybe<Scalars['Float']>;
  gold_price_display?: Maybe<Scalars['Int']>;
  platinum_price_display?: Maybe<Scalars['Int']>;
  founders_price_display?: Maybe<Scalars['Int']>;
};